refactor personal profile component to status indicators
