+++ packages/nextjs/components/StatusIndicators.tsx
@@ -1,8 +1,12 @@
+import { useEffect, useState } from "react";
+import Link from "next/link";
 import { useAccount } from "wagmi";
+import { UserCircleIcon } from "@heroicons/react/24/solid";
 import { useScaffoldReadContract } from "~~/hooks/scaffold-eth";
 
 export const StatusIndicators = () => {
   const { address } = useAccount();
+  const [hasProfile, setHasProfile] = useState(false);
 
   const { data: isBatchMember, isLoading: isLoadingMember } = useScaffoldReadContract({
     contractName: "BatchRegistry",
@@ -15,6 +19,39 @@ export const StatusIndicators = () => {
     functionName: "yourContractAddress",
     args: [address],
   });
+
+  useEffect(() => {
+    const checkProfile = async () => {
+      if (!address) {
+        setHasProfile(false);
+        return;
+      }
+
+      try {
+        const response = await fetch(`/api/check-profile-exists?address=${address}`);
+
+        if (!response.ok) {
+          console.error("Error checking profile - HTTP error:", response.status);
+          setHasProfile(false);
+          return;
+        }
+
+        const data = await response.json();
+
+        if (data && typeof data.exists === "boolean") {
+          setHasProfile(data.exists);
+        } else {
+          console.error("Invalid response format:", data);
+          setHasProfile(false);
+        }
+      } catch (error) {
+        console.error("Error checking profile:", error);
+        setHasProfile(false);
+      }
+    };
+
+    checkProfile();
+  }, [address]);
 
   const isCheckedIn =
     checkedInContractAddress && checkedInContractAddress !== "0x0000000000000000000000000000000000000000";
@@ -77,6 +114,16 @@ export const StatusIndicators = () => {
           </div>
         </div>
       )}
+      {isCheckedIn &&
+        (hasProfile ? (
+          <Link href={`/builders/${address}`}>
+            <UserCircleIcon className="h-5 w-5 text-gray-500 hover:text-secondary cursor-pointer" />
+          </Link>
+        ) : (
+          <div className="tooltip tooltip-bottom" data-tip="No builder page yet">
+            <UserCircleIcon className={`h-6 w-6 text-gray-300 cursor-not-allowed`} />
+          </div>
+        ))}
     </div>
   );
 };
